+++ src/minimal_overlay/bundles/adopt_openjdk/02_install.sh
@@ -6,9 +6,9 @@ set -e
 . ../../common.sh
 
 cd $OVERLAY_WORK_DIR/$BUNDLE_NAME
-mv `ls -d *` adoptopenjdk
+mv * adoptopenjdk
 
-rm -rf $DEST_DIR
+rm -rf "${DEST_DIR:?}"
 mkdir -p $DEST_DIR/bin
 mkdir -p $DEST_DIR/opt
 
@@ -16,9 +16,10 @@ mv adoptopenjdk $DEST_DIR/opt
 
 cd $DEST_DIR/bin
 
-for FILE in $(ls ../opt/adoptopenjdk/bin)
+for FILE in ../opt/adoptopenjdk/bin/*
 do
-  ln -s ../opt/adoptopenjdk/bin/$FILE $FILE
+  [ -e "$FILE" ] || continue
+  ln -s "$FILE" "$(basename "$FILE")"
 done
 
 install_to_overlay
