+++ src/minimal_overlay/bundles/bosh_cli/01_get.sh
@@ -25,7 +25,7 @@ fi
 
 # 删除先前准备好的 BOSH CLI 目录。
 echo "正在清理 BOSH CLI 的工作目录，这可能需要一些时间。"
-rm -rf $WORK_DIR/overlay/$BUNDLE_NAME
+rm -rf "${WORK_DIR:?}/overlay/${BUNDLE_NAME:?}"
 mkdir $WORK_DIR/overlay/$BUNDLE_NAME
 
 # 复制 bosh-cli 到目录 'work/overlay/bosh_cli'。
