+++ src/minimal_overlay/bundles/cf_cli/01_get.sh
@@ -25,7 +25,7 @@ fi
 
 # 删除先前准备好的 cloud foundry cli 目录。
 echo "正在清理 Cloud Foundry CLI 的工作目录，这可能需要一些时间。"
-rm -rf $WORK_DIR/overlay/$BUNDLE_NAME
+rm -rf "${WORK_DIR:?}/overlay/${BUNDLE_NAME:?}"
 mkdir $WORK_DIR/overlay/$BUNDLE_NAME
 
 # 复制 cf-cli.tgz 到目录 'work/overlay/cf_cli'。
