+++ src/minimal_overlay/bundles/kexec_tools/01_get.sh
@@ -28,7 +28,7 @@ fi
 
 # 删除先前解压出的 kexec-tools 目录。
 echo "正在清理 kexec-tools 的工作目录，这可能需要一些时间。"
-rm -rf $WORK_DIR/overlay/$BUNDLE_NAME
+rm -rf "${WORK_DIR:?}/overlay/${BUNDLE_NAME:?}"
 mkdir $WORK_DIR/overlay/$BUNDLE_NAME
 
 # 解压 kexec-tools 到目录 'work/overlay/kexec-tools'。
